pyproject.toml
[project]
name = "py-sandra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "streamlit>=1.30.0",
    "cassandra-driver>=3.28.0",
    "pandas>=2.0.0"
]
